Count the number of recording series for each year in an fhx object

Count the number of recording series for each year in an fhx object

yearly_recording(x, injury_event = FALSE)

Arguments

x

An fhx object.

injury_event

Boolean indicating whether injuries should be considered events. Default is FALSE.

Value

A data frame with columns giving the year and recording events count.

Examples

data(lgr2)
yearly_recording(lgr2)
#>     year Freq
#> 1   1423    2
#> 2   1424    2
#> 3   1425    2
#> 4   1426    2
#> 5   1427    2
#> 6   1428    2
#> 7   1429    2
#> 8   1430    2
#> 9   1431    2
#> 10  1432    2
#> 11  1433    2
#> 12  1434    2
#> 13  1435    2
#> 14  1436    2
#> 15  1437    2
#> 16  1438    2
#> 17  1439    2
#> 18  1440    2
#> 19  1441    2
#> 20  1442    2
#> 21  1443    2
#> 22  1444    2
#> 23  1445    2
#> 24  1446    2
#> 25  1447    2
#> 26  1448    2
#> 27  1449    2
#> 28  1450    2
#> 29  1451    2
#> 30  1452    2
#> 31  1453    2
#> 32  1454    2
#> 33  1455    2
#> 34  1456    2
#> 35  1457    2
#> 36  1458    2
#> 37  1459    2
#> 38  1460    2
#> 39  1461    2
#> 40  1462    2
#> 41  1463    2
#> 42  1464    2
#> 43  1465    2
#> 44  1466    2
#> 45  1467    2
#> 46  1468    2
#> 47  1469    2
#> 48  1470    2
#> 49  1471    2
#> 50  1472    2
#> 51  1473    2
#> 52  1474    2
#> 53  1475    2
#> 54  1476    2
#> 55  1477    2
#> 56  1478    2
#> 57  1479    2
#> 58  1480    2
#> 59  1481    2
#> 60  1482    2
#> 61  1483    1
#> 62  1484    1
#> 63  1485    1
#> 64  1486    1
#> 65  1487    1
#> 66  1488    1
#> 67  1489    1
#> 68  1490    1
#> 69  1491    1
#> 70  1492    1
#> 71  1493    1
#> 72  1494    1
#> 73  1495    1
#> 74  1496    1
#> 75  1497    1
#> 76  1498    1
#> 77  1499    1
#> 78  1500    1
#> 79  1501    1
#> 80  1502    1
#> 81  1503    1
#> 82  1580    1
#> 83  1581    1
#> 84  1582    1
#> 85  1583    1
#> 86  1584    1
#> 87  1585    1
#> 88  1586    1
#> 89  1587    1
#> 90  1588    1
#> 91  1589    1
#> 92  1590    1
#> 93  1591    1
#> 94  1592    1
#> 95  1593    1
#> 96  1594    1
#> 97  1595    1
#> 98  1596    1
#> 99  1597    1
#> 100 1598    1
#> 101 1599    1
#> 102 1600    1
#> 103 1601    1
#> 104 1602    1
#> 105 1603    1
#> 106 1604    1
#> 107 1605    1
#> 108 1606    1
#> 109 1607    1
#> 110 1608    1
#> 111 1609    1
#> 112 1610    1
#> 113 1611    1
#> 114 1612    1
#> 115 1613    1
#> 116 1614    1
#> 117 1615    1
#> 118 1616    1
#> 119 1617    1
#> 120 1618    1
#> 121 1619    1
#> 122 1620    1
#> 123 1621    1
#> 124 1622    1
#> 125 1623    1
#> 126 1624    1
#> 127 1625    1
#> 128 1626    1
#> 129 1627    1
#> 130 1628    1
#> 131 1629    1
#> 132 1630    1
#> 133 1631    1
#> 134 1632    1
#> 135 1633    1
#> 136 1634    1
#> 137 1635    1
#> 138 1636    1
#> 139 1637    1
#> 140 1638    1
#> 141 1639    1
#> 142 1640    1
#> 143 1641    1
#> 144 1642    1
#> 145 1643    1
#> 146 1644    1
#> 147 1645    1
#> 148 1646    1
#> 149 1647    1
#> 150 1648    1
#> 151 1649    1
#> 152 1650    1
#> 153 1651    1
#> 154 1652    1
#> 155 1653    1
#> 156 1654    1
#> 157 1655    1
#> 158 1656    1
#> 159 1657    1
#> 160 1658    1
#> 161 1659    1
#> 162 1660    1
#> 163 1661    1
#> 164 1662    1
#> 165 1663    1
#> 166 1664    1
#> 167 1665    1
#> 168 1666    1
#> 169 1667    1
#> 170 1668    1
#> 171 1669    1
#> 172 1670    1
#> 173 1671    1
#> 174 1672    1
#> 175 1673    1
#> 176 1674    1
#> 177 1675    1
#> 178 1676    1
#> 179 1677    1
#> 180 1678    1
#> 181 1679    1
#> 182 1680    1
#> 183 1681    1
#> 184 1682    1
#> 185 1683    1
#> 186 1684    1
#> 187 1685    1
#> 188 1686    1
#> 189 1687    1
#> 190 1688    1
#> 191 1689    1
#> 192 1690    1
#> 193 1691    1
#> 194 1692    1
#> 195 1693    1
#> 196 1694    1
#> 197 1695    1
#> 198 1696    1
#> 199 1697    1
#> 200 1698    1
#> 201 1699    1
#> 202 1700    1
#> 203 1701    1
#> 204 1702    1
#> 205 1703    1
#> 206 1704    1
#> 207 1705    1
#> 208 1706    1
#> 209 1707    1
#> 210 1708    1
#> 211 1709    1
#> 212 1710    1
#> 213 1711    1
#> 214 1712    1
#> 215 1713    1
#> 216 1714    1
#> 217 1715    1
#> 218 1716    1
#> 219 1717    1
#> 220 1718    1
#> 221 1719    1
#> 222 1720    1
#> 223 1721    1
#> 224 1722    1
#> 225 1723    1
#> 226 1724    1
#> 227 1725    1
#> 228 1726    1
#> 229 1727    1
#> 230 1728    1
#> 231 1729    1
#> 232 1730    1
#> 233 1731    1
#> 234 1732    1
#> 235 1733    1
#> 236 1734    1
#> 237 1735    1
#> 238 1736    1
#> 239 1737    1
#> 240 1738    1
#> 241 1739    1
#> 242 1740    1
#> 243 1741    1
#> 244 1742    1
#> 245 1743    1
#> 246 1744    1
#> 247 1745    1
#> 248 1746    1
#> 249 1747    1
#> 250 1748    1
#> 251 1749    1
#> 252 1750    1
#> 253 1751    1
#> 254 1752    2
#> 255 1753    2
#> 256 1754    2
#> 257 1755    2
#> 258 1756    2
#> 259 1757    2
#> 260 1758    2
#> 261 1759    3
#> 262 1760    3
#> 263 1761    3
#> 264 1762    3
#> 265 1763    3
#> 266 1764    3
#> 267 1765    3
#> 268 1766    3
#> 269 1767    3
#> 270 1768    3
#> 271 1769    3
#> 272 1770    3
#> 273 1771    2
#> 274 1772    2
#> 275 1773    4
#> 276 1774    4
#> 277 1775    4
#> 278 1776    4
#> 279 1777    4
#> 280 1778    4
#> 281 1779    4
#> 282 1780    4
#> 283 1781    4
#> 284 1782    4
#> 285 1783    4
#> 286 1784    4
#> 287 1785    4
#> 288 1786    4
#> 289 1787    4
#> 290 1788    4
#> 291 1789    4
#> 292 1790    4
#> 293 1791    4
#> 294 1792    4
#> 295 1793    4
#> 296 1794    4
#> 297 1795    4
#> 298 1796    4
#> 299 1797    4
#> 300 1798    4
#> 301 1799    4
#> 302 1800    4
#> 303 1801    4
#> 304 1802    4
#> 305 1803    4
#> 306 1804    4
#> 307 1805    3
#> 308 1806    8
#> 309 1807    8
#> 310 1808    8
#> 311 1809    8
#> 312 1810    8
#> 313 1811    8
#> 314 1812    7
#> 315 1813    7
#> 316 1814    7
#> 317 1815    7
#> 318 1816    7
#> 319 1817    7
#> 320 1818    7
#> 321 1819    7
#> 322 1820    7
#> 323 1821    7
#> 324 1822    7
#> 325 1823    7
#> 326 1824    7
#> 327 1825    7
#> 328 1826    7
#> 329 1827    6
#> 330 1828    6
#> 331 1829    6
#> 332 1830    6
#> 333 1831    6
#> 334 1832    6
#> 335 1833    6
#> 336 1834    6
#> 337 1835    6
#> 338 1836    6
#> 339 1837    6
#> 340 1838    5
#> 341 1839    5
#> 342 1840    5
#> 343 1841    5
#> 344 1842    5
#> 345 1843    5
#> 346 1844    4
#> 347 1845    4
#> 348 1846    4
#> 349 1847    4
#> 350 1848    4
#> 351 1849    4
#> 352 1850    3
#> 353 1851    3
#> 354 1852    2
#> 355 1853    2
#> 356 1854    1
#> 357 1855    1
#> 358 1856    1
#> 359 1857    1
#> 360 1858    1
#> 361 1859    1
#> 362 1860    1
#> 363 1861    2
#> 364 1862    2
#> 365 1863    2
#> 366 1864    2
#> 367 1865    3
#> 368 1866    3
#> 369 1867    4
#> 370 1868    6
#> 371 1869    7
#> 372 1870    7
#> 373 1871    7
#> 374 1872    7
#> 375 1873    7
#> 376 1874    7
#> 377 1875    8
#> 378 1876    9
#> 379 1877   10
#> 380 1878   11
#> 381 1879   11
#> 382 1880   11
#> 383 1881   11
#> 384 1882   11
#> 385 1883   11
#> 386 1884   12
#> 387 1885   12
#> 388 1886   12
#> 389 1887   12
#> 390 1888   12
#> 391 1889   12
#> 392 1890   12
#> 393 1891   12
#> 394 1892   12
#> 395 1893   12
#> 396 1894   13
#> 397 1895   13
#> 398 1896   13
#> 399 1897   13
#> 400 1898   13
#> 401 1899   13
#> 402 1900   13
#> 403 1901   13
#> 404 1902   13
#> 405 1903   13
#> 406 1904   14
#> 407 1905   14
#> 408 1906   14
#> 409 1907   14
#> 410 1908   14
#> 411 1909   14
#> 412 1910   14
#> 413 1911   14
#> 414 1912   14
#> 415 1913   14
#> 416 1914   14
#> 417 1915   14
#> 418 1916   14
#> 419 1917   14
#> 420 1918   14
#> 421 1919   14
#> 422 1920   14
#> 423 1921   14
#> 424 1922   14
#> 425 1923   14
#> 426 1924   14
#> 427 1925   14
#> 428 1926   14
#> 429 1927   14
#> 430 1928   14
#> 431 1929   14
#> 432 1930   14
#> 433 1931   14
#> 434 1932   14
#> 435 1933   14
#> 436 1934   14
#> 437 1935   14
#> 438 1936   14
#> 439 1937   14
#> 440 1938   14
#> 441 1939   14
#> 442 1940   14
#> 443 1941   14
#> 444 1942   14
#> 445 1943   14
#> 446 1944   14
#> 447 1945   14
#> 448 1946   14
#> 449 1947   14
#> 450 1948   14
#> 451 1949   14
#> 452 1950   14
#> 453 1951   14
#> 454 1952   14
#> 455 1953   14
#> 456 1954   14
#> 457 1955   14
#> 458 1956   14
#> 459 1957   14
#> 460 1958   14
#> 461 1959   14
#> 462 1960   14
#> 463 1961   14
#> 464 1962   14
#> 465 1963   14
#> 466 1964   14
#> 467 1965   14
#> 468 1966   14
#> 469 1967   14
#> 470 1968   14
#> 471 1969   14
#> 472 1970   14
#> 473 1971   13
#> 474 1972   13
#> 475 1973   13
#> 476 1974   13
#> 477 1975   13
#> 478 1976   13
#> 479 1977   13
#> 480 1978   13
#> 481 1979   13
#> 482 1980   13
#> 483 1981   13
#> 484 1982   13
#> 485 1983   13
#> 486 1984   13
#> 487 1985   13
#> 488 1986   13
#> 489 1987   13
#> 490 1988   13
#> 491 1989   13
#> 492 1990   13
#> 493 1991   13
#> 494 1992   13
#> 495 1993   13
#> 496 1994   13
#> 497 1995   13
#> 498 1996   13
#> 499 1997   13
#> 500 1998   13
#> 501 1999   13
#> 502 2000   13
#> 503 2001   13
#> 504 2002   13
#> 505 2003   13
#> 506 2004   13
#> 507 2005   13
#> 508 2006   13
#> 509 2007   13
#> 510 2008   13
#> 511 2009   13
#> 512 2010   13
#> 513 2011   13
#> 514 2012    2