Get years with events for an fhx object

Get years with events for an fhx object

get_event_years(
  x,
  scar_event = TRUE,
  injury_event = FALSE,
  custom_grep_str = NULL
)

Arguments

x
An fhx object.

scar_event
Boolean indicating whether years with scar events should be returned. Default is TRUE.

injury_event
Boolean indicating whether years with injury events should be returned. Default is FALSE.

custom_grep_str
Character string to pass a custom grep search pattern to search x "rec_type" column for. NULL by default.

Value

A list. Elements of the list are numeric vectors giving the years with events for each fhx series. Each element's name reflects the series' name.

See also

Examples

data(pgm)
get_event_years(pgm, scar_event = TRUE, injury_event = TRUE)

#> $PGM45
#> integer(0)
#> 
#> $PGM44
#> integer(0)
#> 
#> $PGM43
#> integer(0)
#> 
#> $PGM42
#> integer(0)
#> 
#> $PGM38
#> integer(0)
#> 
#> $PGM37
#> integer(0)
#> 
#> $PGM34
#> integer(0)
#> 
#> $PGM33
#> integer(0)
#> 
#> $PGM30
#> integer(0)
#> 
#> $PGM29
#> integer(0)
#> 
#> $PGM28
#> integer(0)
#> 
#> $PGM27
#> integer(0)
#> 
#> $PGM22
#> integer(0)
#> 
#> $PGM21
#> integer(0)
#> 
#> $PGM20
#> integer(0)
#> 
#> $PGM19
#> integer(0)
#> 
#> $PGM18
#> integer(0)
#> 
#> $PGM16
#> integer(0)
#> 
#> $PGM14
#> integer(0)
#> 
#> $PGM13
#> integer(0)
#> 
#> $PGM2
#> integer(0)
#> 
#> $PGM15
#> integer(0)
#> 
#> $PGM5
#> integer(0)
#> 
#> $PGM40
#> integer(0)
#> 
#> $PGM36
#> integer(0)
#> 
#> $PGM3
#> integer(0)
#> 
#> $PGM17
#> integer(0)
#> 
#> $PGM11
#> integer(0)
#> 
#> $PGM7
#> [1] 1993
#> 
#> $PGM23
#> [1] 1993
#> 
#> $PGM25
#> [1] 1993
#> 
#> $PGM10
#> [1] 1993
#> 
#> $PGM6
#> [1] 1993
#> 
#> $PGM1
#> [1] 1993
#> 
#> $PGM26
#> [1] 1993
#> 
#> $PGM9
#> [1] 1993
#> 
#> $PGM39
#> [1] 1798 1818 1824 1842 1863
#> 
#> $PGM35
#> [1] 1737 1748 1789 1798
#> 
#> $PGM12
#> [1] 1664 1782 1795 1801 1806 1818 1829 1863 1871
#> 
#> $PGM31
#>  [1] 1625 1636 1648 1664 1674 1685 1714 1724 1752 1782 1824
#> 
#> $PGM32
#>  [1] 1636 1664 1724 1746 1752 1789 1798 1806 1824 1842 1863 1871 1993
#> 
#> attr(,"split_type")
#> [1] "data.frame"
#> attr(,"split_labels")
#>    series
#> 1   PGM45
#> 2   PGM44
#> 3   PGM43
#> 4   PGM42
#> 5   PGM38
#> 6   PGM37
#> 7   PGM34
#> 8   PGM33
#> 9   PGM30
#> 10  PGM29
#> 11  PGM28
#> 12  PGM27
#> 13  PGM22
#> 14  PGM21
#> 15  PGM20
#> 16  PGM19
#> 17  PGM18
#> 18  PGM16
#> 19  PGM14
#> 20  PGM13
#> 21   PGM2
#> 22  PGM15
#> 23   PGM5
#> 24  PGM40
#> 25  PGM36
#> 26   PGM3
#> 27  PGM17
#> 28  PGM11
#> 29   PGM7
#> 30  PGM23
#> 31  PGM25
#> 32  PGM10
#> 33   PGM6
#> 34   PGM1
#> 35  PGM26
#> 36   PGM9
#> 37  PGM39
#> 38  PGM35
#> 39  PGM12
#> 40  PGM31
#> 41  PGM32

# Passing a custom string to grep. This one identified recorder years:
get_event_years(pgm, custom_grep_str = "recorder_")

#> burnr::get_events(): custom_search_str was defined, ignoring scar_event and injury_event arguments

#> $PGM45
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM44
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM43
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM42
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM38
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM37
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM34
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM33
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM30
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM29
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM28
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM27
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM22
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM21
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM20
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM19
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM18
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM16
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM14
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM13
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM2
#>  [1] 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986
#> [16] 1987 1988 1989 1990 1991
#> 
#> $PGM15
#>  [1] 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949 1950 1951 1952 1953 1954
#> [16] 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969
#> [31] 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984
#> [46] 1985 1986 1987 1988 1989 1990 1991
#> 
#> $PGM5
#>  [1] 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949 1950 1951 1952
#> [16] 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967
#> [31] 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982
#> [46] 1983 1984 1985 1986 1987 1988 1989 1990 1991
#> 
#> $PGM40
#>  [1] 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948
#> [16] 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963
#> [31] 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978
#> [46] 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991
#> 
#> $PGM36
#>  [1] 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930
#> [16] 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945
#> [31] 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960
#> [46] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#> [61] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#> [76] 1991
#> 
#> $PGM3
#>  [1] 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915 1916 1917 1918 1919 1920
#> [16] 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932 1933 1934 1935
#> [31] 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949 1950
#> [46] 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965
#> [61] 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980
#> [76] 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991
#> 
#> $PGM17
#>  [1] 1901 1902 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915
#> [16] 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930
#> [31] 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945
#> [46] 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960
#> [61] 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975
#> [76] 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990
#> [91] 1991
#> 
#> $PGM11
#>   [1] 1817 1818 1819 1820 1821 1822 1823 1824 1825 1826 1827 1828 1829 1830 1831
#>  [16] 1832 1833 1834 1835 1836 1837 1838 1839 1840 1841 1842 1843 1844 1845 1846
#>  [31] 1847 1848 1849 1850 1851 1852 1853 1854 1855 1856 1857 1858 1859 1860 1861
#>  [46] 1862 1863 1864 1865 1866 1867 1868 1869 1870 1871 1872 1873 1874 1875 1876
#>  [61] 1877 1878 1879 1880 1881 1882 1883 1884 1885 1886 1887 1888 1889 1890 1891
#>  [76] 1892 1893 1894 1895 1896 1897 1898 1899 1900 1901 1902 1903 1904 1905 1906
#>  [91] 1907 1908 1909 1910 1911 1912 1913 1914 1915 1916 1917 1918 1919 1920 1921
#> [106] 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932 1933 1934 1935 1936
#> [121] 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949 1950 1951
#> [136] 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964 1965 1966
#> [151] 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979 1980 1981
#> [166] 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991
#> 
#> $PGM7
#>  [1] 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932 1933 1934
#> [16] 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947 1948 1949
#> [31] 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962 1963 1964
#> [46] 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977 1978 1979
#> [61] 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992 1994 1995
#> [76] 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008 2009 2010
#> [91] 2011 2012
#> 
#> $PGM23
#>  [1] 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932
#> [16] 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947
#> [31] 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962
#> [46] 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977
#> [61] 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992
#> [76] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [91] 2009 2010 2011
#> 
#> $PGM25
#>  [1] 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931
#> [16] 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946
#> [31] 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961
#> [46] 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976
#> [61] 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991
#> [76] 1992 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007
#> [91] 2008 2009 2010 2011 2012
#> 
#> $PGM10
#>   [1] 1883 1884 1885 1886 1887 1888 1889 1890 1891 1892 1893 1894 1895 1896 1897
#>  [16] 1898 1899 1900 1901 1902 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912
#>  [31] 1913 1914 1915 1916 1917 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927
#>  [46] 1928 1929 1930 1931 1932 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942
#>  [61] 1943 1944 1945 1946 1947 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957
#>  [76] 1958 1959 1960 1961 1962 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972
#>  [91] 1973 1974 1975 1976 1977 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987
#> [106] 1988 1989 1990 1991 1992 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003
#> [121] 2004 2005 2006 2007 2008 2009 2010 2011 2012
#> 
#> $PGM6
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM1
#> [1] 1994 1995 1996 1997 1998
#> 
#> $PGM26
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM9
#>  [1] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [16] 2009 2010 2011 2012
#> 
#> $PGM39
#>  [1] 1799 1800 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 1811 1812 1813
#> [16] 1814 1815 1816 1817 1819 1820 1821 1822 1823 1825 1826 1827 1828 1829 1830
#> [31] 1831 1832 1833 1834 1835 1836 1837 1838 1839 1840 1841 1843 1844 1845 1846
#> [46] 1847 1848 1849 1850 1851 1852 1853 1854 1855 1856 1857 1858 1859 1860 1861
#> [61] 1862 1864 1865 1866 1867 1868 1869 1870 1871
#> 
#> $PGM35
#>   [1] 1738 1739 1740 1741 1742 1743 1744 1745 1746 1747 1749 1750 1751 1752 1753
#>  [16] 1754 1755 1756 1757 1758 1759 1760 1761 1762 1763 1764 1765 1766 1767 1768
#>  [31] 1769 1770 1771 1772 1773 1774 1775 1776 1777 1778 1779 1780 1781 1782 1783
#>  [46] 1784 1785 1786 1787 1788 1790 1791 1792 1793 1794 1795 1796 1797 1799 1800
#>  [61] 1801 1802 1803 1804 1805 1806 1807 1808 1809 1810 1811 1812 1813 1814 1815
#>  [76] 1816 1817 1818 1819 1820 1821 1822 1823 1824 1825 1826 1827 1828 1829 1830
#>  [91] 1831 1832 1833 1834 1835 1836 1837 1838 1839 1840 1841 1842 1843
#> 
#> $PGM12
#>  [1] 1783 1784 1785 1786 1787 1788 1789 1790 1791 1792 1793 1794 1796 1797 1798
#> [16] 1799 1800 1802 1803 1804 1805 1807 1808 1809 1810 1811 1812 1813 1814 1815
#> [31] 1816 1817 1819 1820 1821 1822 1823 1824 1825 1826 1827 1828 1830 1831 1832
#> [46] 1833 1834 1835 1836 1837 1838 1839 1840 1841 1842 1843 1844 1845 1846 1847
#> [61] 1848 1849 1850 1851 1852 1853 1854 1855 1856 1857 1858 1859 1860 1861 1862
#> [76] 1864 1865 1866 1867 1868 1869 1870 1872 1873 1874 1875
#> 
#> $PGM31
#>   [1] 1637 1638 1639 1640 1641 1642 1643 1644 1645 1646 1647 1649 1650 1651 1652
#>  [16] 1653 1654 1655 1656 1657 1658 1659 1660 1661 1662 1663 1665 1666 1667 1668
#>  [31] 1669 1670 1671 1672 1673 1675 1676 1677 1678 1679 1680 1681 1682 1683 1684
#>  [46] 1686 1687 1688 1689 1690 1691 1692 1693 1694 1695 1696 1697 1698 1699 1700
#>  [61] 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710 1711 1712 1713 1715 1716
#>  [76] 1717 1718 1719 1720 1721 1722 1723 1725 1726 1727 1728 1729 1730 1731 1732
#>  [91] 1733 1734 1735 1736 1737 1738 1739 1740 1741 1742 1743 1744 1745 1746 1747
#> [106] 1748 1749 1750 1751 1753 1754 1755 1756 1757 1758 1759 1760 1761 1762 1763
#> [121] 1764 1765 1766 1767 1768 1769 1770 1771 1772 1773 1774 1775 1776 1777 1778
#> [136] 1779 1780 1781 1783 1784 1785 1786 1787 1788 1789 1790 1791 1792 1793 1794
#> [151] 1795 1796 1797 1798 1799 1800 1801 1802 1803 1804 1805 1806 1807 1808 1809
#> [166] 1810 1811 1812 1813 1814 1815 1816 1817 1818 1819 1820 1821 1822 1823
#> 
#> $PGM32
#>   [1] 1637 1638 1639 1640 1641 1642 1643 1644 1645 1646 1647 1648 1649 1650 1651
#>  [16] 1652 1653 1654 1655 1656 1657 1658 1659 1660 1661 1662 1663 1665 1666 1667
#>  [31] 1668 1669 1670 1671 1672 1673 1674 1675 1676 1677 1678 1679 1680 1681 1682
#>  [46] 1683 1684 1685 1686 1687 1688 1689 1690 1691 1692 1693 1694 1695 1696 1697
#>  [61] 1698 1699 1700 1701 1702 1703 1704 1705 1706 1707 1708 1709 1710 1711 1712
#>  [76] 1713 1714 1715 1716 1717 1718 1719 1720 1721 1722 1723 1725 1726 1727 1728
#>  [91] 1729 1730 1731 1732 1733 1734 1735 1736 1737 1738 1739 1740 1741 1742 1743
#> [106] 1744 1745 1747 1748 1749 1750 1751 1753 1754 1755 1756 1757 1758 1759 1760
#> [121] 1761 1762 1763 1764 1765 1766 1767 1768 1769 1770 1771 1772 1773 1774 1775
#> [136] 1776 1777 1778 1779 1780 1781 1782 1783 1784 1785 1786 1787 1788 1790 1791
#> [151] 1792 1793 1794 1795 1796 1797 1799 1800 1801 1802 1803 1804 1805 1807 1808
#> [166] 1809 1810 1811 1812 1813 1814 1815 1816 1817 1818 1819 1820 1821 1822 1823
#> [181] 1825 1826 1827 1828 1829 1830 1831 1832 1833 1834 1835 1836 1837 1838 1839
#> [196] 1840 1841 1843 1844 1845 1846 1847 1848 1849 1850 1851 1852 1853 1854 1855
#> [211] 1856 1857 1858 1859 1860 1861 1862 1864 1865 1866 1867 1868 1869 1870 1872
#> [226] 1873 1874 1875 1876 1877 1878 1879 1880 1881 1882 1883 1884 1885 1886 1887
#> [241] 1888 1889 1890 1891 1892 1893 1894 1895 1896 1897 1898 1899 1900 1901 1902
#> [256] 1903 1904 1905 1906 1907 1908 1909 1910 1911 1912 1913 1914 1915 1916 1917
#> [271] 1918 1919 1920 1921 1922 1923 1924 1925 1926 1927 1928 1929 1930 1931 1932
#> [286] 1933 1934 1935 1936 1937 1938 1939 1940 1941 1942 1943 1944 1945 1946 1947
#> [301] 1948 1949 1950 1951 1952 1953 1954 1955 1956 1957 1958 1959 1960 1961 1962
#> [316] 1963 1964 1965 1966 1967 1968 1969 1970 1971 1972 1973 1974 1975 1976 1977
#> [331] 1978 1979 1980 1981 1982 1983 1984 1985 1986 1987 1988 1989 1990 1991 1992
#> [346] 1994 1995 1996 1997 1998 1999 2000 2001 2002 2003 2004 2005 2006 2007 2008
#> [361] 2009 2010 2011 2012
#> 
#> attr(,"split_type")
#> [1] "data.frame"
#> attr(,"split_labels")
#>    series
#> 1   PGM45
#> 2   PGM44
#> 3   PGM43
#> 4   PGM42
#> 5   PGM38
#> 6   PGM37
#> 7   PGM34
#> 8   PGM33
#> 9   PGM30
#> 10  PGM29
#> 11  PGM28
#> 12  PGM27
#> 13  PGM22
#> 14  PGM21
#> 15  PGM20
#> 16  PGM19
#> 17  PGM18
#> 18  PGM16
#> 19  PGM14
#> 20  PGM13
#> 21   PGM2
#> 22  PGM15
#> 23   PGM5
#> 24  PGM40
#> 25  PGM36
#> 26   PGM3
#> 27  PGM17
#> 28  PGM11
#> 29   PGM7
#> 30  PGM23
#> 31  PGM25
#> 32  PGM10
#> 33   PGM6
#> 34   PGM1
#> 35  PGM26
#> 36   PGM9
#> 37  PGM39
#> 38  PGM35
#> 39  PGM12
#> 40  PGM31
#> 41  PGM32

# Use with composite to get composite years:
comp <- composite(pgm, comp_name = "pgm")
event_yrs <- get_event_years(comp)[["pgm"]]
print(event_yrs)

#>  [1] 1636 1648 1664 1674 1685 1714 1724 1737 1746 1748 1752 1782 1789 1795 1798
#> [16] 1806 1818 1824 1842 1863 1871